scarlet_queen_core::each_individual

Trait ReplenisherIndividualTrait

pub trait ReplenisherIndividualTrait<const N: usize, const R: usize>: EachCrateIndividual {
    // Required method
    fn replenish<'a, G>(group: G) -> Vec<<Self as EachCrateIndividual>::Item>
       where G: IntoIterator<Item = &'a Self>,
             Self: 'a;
}

A trait for individual defined by replenisher crate.

T - A type of value.

N - The number of individuals.

R - The number of individuals after individuals are reduced by selector.

Required Methods

fn replenish<'a, G>(group: G) -> Vec<<Self as EachCrateIndividual>::Item>
where G: IntoIterator<Item = &'a Self>, Self: 'a,

Replenish individuals.

'a - A lifetime of group.

G - A type of group.

group - A value which you are able to get Self from.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.