scarlet_queen_core::pokemon_type

Trait PokemonTypeTrait

pub trait PokemonTypeTrait:
    Into<PokemonTypeAll>
    + TryFrom<PokemonTypeAll>
    + Clone
    + Eq
    + Hash {
    const ALL_LEN: usize;
    const ALL: [Option<Self>; 19];

    // Provided methods
    fn sample<R>(rng: &mut R) -> Self
       where R: Rng + Sized { ... }
    fn color_map(&self) -> impl Color { ... }
}

A trait for a enum which is a PokemonTypeAll subset.

Example

use scarlet_queen_core::pokemon_type::{PokemonTypeTrait, PokemonTypeAll};
use plotters::style::Color;

#[derive(Clone, PartialEq, Eq, Hash)]
enum PTTraitSample {
    Normal,
    Water,
}
impl Into<PokemonTypeAll> for PTTraitSample {
    fn into(self) -> PokemonTypeAll {
        match self {
            PTTraitSample::Normal => PokemonTypeAll::Normal,
            PTTraitSample::Water => PokemonTypeAll::Water,
        }
    }
}
impl TryFrom<PokemonTypeAll> for PTTraitSample {
    type Error = ();
    fn try_from(value: PokemonTypeAll) -> Result<Self, Self::Error> {
        match value {
            PokemonTypeAll::Normal => Ok(PTTraitSample::Normal),
            PokemonTypeAll::Water => Ok(PTTraitSample::Water),
            _ => Err(()),
        }
    }
}
impl PokemonTypeTrait for PTTraitSample {
    const ALL_LEN: usize = 2;
    const ALL: [Option<Self>; 19] = [
        Some(PTTraitSample::Normal),
        Some(PTTraitSample::Water),
        None,
        None,
        None,
        None,
        None,
        None,
        None,
        None,
        None,
        None,
        None,
        None,
        None,
        None,
        None,
        None,
        None
    ];
}

let sample: PTTraitSample = PTTraitSample::Water;

assert_eq!(sample.color_map().rgb(), (41, 146, 255));

let mut thread_rng: rand::prelude::ThreadRng = rand::rng();
let sample: PTTraitSample = PTTraitSample::sample(&mut thread_rng);

assert!(
    [
        PTTraitSample::Normal,
        PTTraitSample::Water,
    ]
        .contains(&sample)
)

Required Associated Constants

const ALL_LEN: usize

The size of subset.

const ALL: [Option<Self>; 19]

All of this values.

Provided Methods

fn sample<R>(rng: &mut R) -> Self
where R: Rng + Sized,

Generate a random pokemon type which this type contains.

Panics

An error may be occured if all of Self::ALL[..Self::ALL_LEN] is not Some.

fn color_map(&self) -> impl Color

Get a color of a pokemon type.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.